+++ src/Adafruit.Devices.Veml7700/GainLevel.cs
@@ -1,0 +1,9 @@
+namespace Adafruit.Devices.Veml7700
+{
+    public enum GainLevel : byte { 
+        Level1 = 0,
+        Level2 = 1,
+        Level1_8 = 2,
+        Level1_4 = 3
+    }
+}
