+++ src/Adafruit.Devices.Veml7700/IAdafruit_VEML7700.cs
@@ -4,22 +4,6 @@
 
 namespace Adafruit.Devices.Veml7700
 {
-    public enum GainLevel : byte { 
-        Level1 = 0,
-        Level2 = 1,
-        Level1_8 = 2,
-        Level1_4 = 3
-    }
-
-    public enum IntegrationTime : byte {
-        IT_100MS = 0,
-        IT_200MS = 1,
-        IT_400MS = 2,
-        IT_800MS = 3,
-        IT_50MS = 8,
-        IT_25MS = 12
-    }
-
     /// <summary>
     /// Identifies a device driver for the Adafruit VEML7700.
     /// </summary>
