+++ src/Adafruit.Devices.Veml7700/IntegrationTime.cs
@@ -1,0 +1,11 @@
+namespace Adafruit.Devices.Veml7700
+{
+    public enum IntegrationTime : byte {
+        IT_100MS = 0,
+        IT_200MS = 1,
+        IT_400MS = 2,
+        IT_800MS = 3,
+        IT_50MS = 8,
+        IT_25MS = 12
+    }
+}
